requirements.txt
kafka-python
pyspark==3.4.0
apache-airflow-providers-apache-spark==4.1.0